sudo not required and creates complications
This script was designed to be run by a non privileged user.  A user doesn't need sudo to edit their own crontab.
sudo prompts for a password, so if cron runs this in the background after a successful execution of /usr/local/bin/noip-renew-myusername.sh, it will wait for a password and never update crontab.